Cache progress bar output to speed it up
* If a progress bar is requested that looks exactly like the currently
displayed progress bar, do not display it. This is faster in cases where
code calls fProgressBar more often then the bar is long. Since the exact
progress bar being requested is already displayed, this does not change
what the user sees.
* Updated progress bar test to call the fProgressBar function a lot more
often so failure to cache as described above would result in very slow
test runs. There is not automatic test for this atm; the user running
the test should notice this and fix the code if needed.